teams_etl/requirements.txt
requests
pandas
google-cloud-bigquery
pandas-gbq
pyarrow
db-dtypes